+++ HISTORY.md
@@ -8,8 +8,10 @@ All notable changes to this project will be documented in this file.
   * Raised dependency `cyclonedx/cyclonedx-library:^1.4.2`, was `cyclonedx/cyclonedx-library:^1.3.1`. (via [#192])
 * Misc
   * Adjusted internal typing and typehints. (via [#192])
+  * Improved compatibility to Composer v2.3 (via [#212])
 
 [#192]: https://github.com/CycloneDX/cyclonedx-php-composer/pull/192
+[#212]: https://github.com/CycloneDX/cyclonedx-php-composer/pull/212
 
 ## 3.9.2 - 2021-12-04